Complete ARCHIVED state and VAL-020.1 specification
ARCHIVED State (run-000006):
- Immutable archive of run-000005-VERIFIED
- Preservation-only: no computation, only storage
- Append-only policy enforced
- Verification hash: 209e1ca5deeaee96...
- Complete lifecycle: DESIGNED → ARCHIVED

VAL-020.1 Specification:
- Transformer Block Verification
- Composed inference graph: 9 stages
- Execution topology: Embedding → RMSNorm → QKV → RoPE → Attention → KV Cache → FFN → Residual
- Acceptance gates: G1-G6 (tensor provenance, dispatch, numerical, determinism, telemetry, seal)
- Evidence format: rawrxd-val-020-v1
- Implementation plan: 6 phases

This establishes the pattern for scaling from primitives
to composed execution graphs with cryptographic evidence.

Next: Implement VAL-020.1 kernels (QKV, RoPE, Attention, KV Cache, FFN)

+++ src/integration/val016_source_repair_v2.cpp
@@ -545,7 +545,7 @@ public:
         std::filesystem::create_directories(basePath_ + "/repair");
         nlohmann::json j;
         j["classification"] = classification.description;
-        j["confidence"] = static_cast<double>(classification.confidence);
+        j["confidence"] = classification.confidence;
         j["reason"] = classification.reason;
         j["patch"] = patch.toJson();
         std::ofstream file(basePath_ + "/repair/repair_plan.json");
